move argument parsing to separate file
